+++ Demo/api/models/usersModel.js
@@ -18,16 +18,16 @@ const usersSchema = new mongoose.Schema({
     type: String,
     required: true,
   },
-  passwordConf: {
-    type: String,
-    required: true,
+  createdAt: {
+    type: Date,
+    default: Date.now
   }
 });
 
 //authenticate input against database
-usersSchema.statics.authenticate = function (email, password, callback) {
-  User.findOne({ email: email })
-    .exec(function (err, user) {
+usersSchema.statics.authenticate = (email, password, callback) => {
+  Users.findOne({ email: email })
+    .exec((err, user) => {
       if (err) {
         return callback(err)
       } else if (!user) {
@@ -35,11 +35,11 @@ usersSchema.statics.authenticate = function (email, password, callback) {
         err.status = 401;
         return callback(err);
       }
-      bcrypt.compare(password, user.password, function (err, result) {
+      bcrypt.compare(password, user.password, (err, result) => {
         if (result === true) {
           return callback(null, user);
         } else {
-          return callback();
+          return callback(err);
         }
       })
     });
@@ -48,7 +48,7 @@ usersSchema.statics.authenticate = function (email, password, callback) {
 //hashing a password before saving it to the database
 usersSchema.pre('save', function (next) {
   var user = this;
-  bcrypt.hash(user.password, 10, function (err, hash) {
+  bcrypt.hash(user.password, 10, (err, hash) => {
     if (err) {
       return next(err);
     }
@@ -57,6 +57,8 @@ usersSchema.pre('save', function (next) {
   })
 });
 
+
+const Users = mongoose.model('Users', usersSchema);
 module.exports = {
-  Users : mongoose.model('Users', usersSchema)
+  Users
 }