+++ CHANGELOG.md
@@ -12,13 +12,19 @@
 
 ### Version 0.0.3
 
- - Auto generate document API. *path /api-docs*
+ - Auto generate document API. *path /api-docs*.
 
 ### Version 0.0.4
 
- - Add authentication with user  `{"email": "String", "username": "String", "password": "String", "passwordConf": "String"}`
+ - Add authentication with user  `{"email": "String", "username": "String", "password": "String"}`/
 
 ### Version 0.0.5
  
  - rename package name.
  - Upload to npmjs.
+
+### Version 0.0.6
+
+ - Update API default.
+ - Add document for authentication.
+ - Update users model.